Creating a TestBase class
